Version 1.2.2
* Added a Makefile that can create deb packages
* Added `build/`/`*.swp`/`.#*` to `.gitignore`
* Added my email to the `author` in `metadata.json`
* Fixed more problems that `clangd` was unhappy about
* Added missing includes (LSP doing something useful for once)
* Removed useless includes
* Updated `README.md`

+++ src/clock.c
@@ -116,7 +116,7 @@ const char* ZERO_FILLS[] = {"0","1","2","3","4","5","6","7","8","9",":"};
 void display_number(RowColPair pos, Character chr, const char* ch0,
                     const char* ch1, uint8_t scale) {
     RowColPair term_size = screen_term_size();
-    uint8_t _mask = 1 << (chr[0]-1);
+    uint8_t _mask = (uint8_t) (1 << (chr[0]-1));
     for (uint8_t r=0; r<chr[1]; r++) {
         for (uint8_t rs=0; rs<scale; rs++) {
             uint8_t row_data = chr[r+2]; // [width height ...data]
